+++ src/config.ts
@@ -12,11 +12,16 @@ import { getTelemetryId } from './redhatTelemetry';
 class Config {
   telemetryId: string;
   stackAnalysisCommand: string;
-  rhRepositoryRecommendationNotificationCommand: string;
+  trackRecommendationAcceptanceCommand: string;
   utmSource: string;
   matchManifestVersions: string;
+  usePythonVirtualEnvironment: string;
+  useGoMVS: string;
+  enablePythonBestEffortsInstallation: string;
+  usePipDepTree: string;
   vulnerabilityAlertSeverity: string;
   exhortMvnPath: string;
+  exhortGradlePath: string;
   exhortNpmPath: string;
   exhortGoPath: string;
   exhortPython3Path: string;
@@ -25,14 +30,26 @@ class Config {
   exhortPipPath: string;
   rhdaReportFilePath: string;
   secrets: vscode.SecretStorage;
+  exhortSyftPath: string;
+  exhortSyftConfigPath: string;
+  exhortSkopeoPath: string;
+  exhortSkopeoConfigPath: string;
+  exhortDockerPath: string;
+  exhortPodmanPath: string;
+  exhortImagePlatform: string;
 
   private readonly DEFAULT_MVN_EXECUTABLE = 'mvn';
+  private readonly DEFAULT_GRADLE_EXECUTABLE = 'gradle';
   private readonly DEFAULT_NPM_EXECUTABLE = 'npm';
   private readonly DEFAULT_GO_EXECUTABLE = 'go';
   private readonly DEFAULT_PYTHON3_EXECUTABLE = 'python3';
   private readonly DEFAULT_PIP3_EXECUTABLE = 'pip3';
   private readonly DEFAULT_PYTHON_EXECUTABLE = 'python';
   private readonly DEFAULT_PIP_EXECUTABLE = 'pip';
+  private readonly DEFAULT_SYFT_EXECUTABLE = 'syft';
+  private readonly DEFAULT_SKOPEO_EXECUTABLE = 'skopeo';
+  private readonly DEFAULT_DOCKER_EXECUTABLE = 'docker';
+  private readonly DEFAULT_PODMAN_EXECUTABLE = 'podman';
 
   /**
    * Creates an instance of the Config class.
@@ -58,20 +75,36 @@ class Config {
     const rhdaConfig = this.getRhdaConfig();
 
     this.stackAnalysisCommand = commands.STACK_ANALYSIS_COMMAND;
-    this.rhRepositoryRecommendationNotificationCommand = commands.REDHAT_REPOSITORY_RECOMMENDATION_NOTIFICATION_COMMAND;
+    this.trackRecommendationAcceptanceCommand = commands.TRACK_RECOMMENDATION_ACCEPTANCE_COMMAND;
     this.utmSource = GlobalState.UTM_SOURCE;
     /* istanbul ignore next */
     this.matchManifestVersions = rhdaConfig.matchManifestVersions ? 'true' : 'false';
+    /* istanbul ignore next */
+    this.usePythonVirtualEnvironment = rhdaConfig.usePythonVirtualEnvironment ? 'true' : 'false';
+    /* istanbul ignore next */
+    this.useGoMVS = rhdaConfig.useGoMVS ? 'true' : 'false';
+    /* istanbul ignore next */
+    this.enablePythonBestEffortsInstallation = rhdaConfig.enablePythonBestEffortsInstallation ? 'true' : 'false';
+    /* istanbul ignore next */
+    this.usePipDepTree = rhdaConfig.usePipDepTree ? 'true' : 'false';
     this.vulnerabilityAlertSeverity = rhdaConfig.vulnerabilityAlertSeverity;
     /* istanbul ignore next */
     this.rhdaReportFilePath = rhdaConfig.reportFilePath || DEFAULT_RHDA_REPORT_FILE_PATH;
     this.exhortMvnPath = rhdaConfig.mvn.executable.path || this.DEFAULT_MVN_EXECUTABLE;
+    this.exhortGradlePath = rhdaConfig.gradle.executable.path || this.DEFAULT_GRADLE_EXECUTABLE;
     this.exhortNpmPath = rhdaConfig.npm.executable.path || this.DEFAULT_NPM_EXECUTABLE;
     this.exhortGoPath = rhdaConfig.go.executable.path || this.DEFAULT_GO_EXECUTABLE;
     this.exhortPython3Path = rhdaConfig.python3.executable.path || this.DEFAULT_PYTHON3_EXECUTABLE;
     this.exhortPip3Path = rhdaConfig.pip3.executable.path || this.DEFAULT_PIP3_EXECUTABLE;
     this.exhortPythonPath = rhdaConfig.python.executable.path || this.DEFAULT_PYTHON_EXECUTABLE;
     this.exhortPipPath = rhdaConfig.pip.executable.path || this.DEFAULT_PIP_EXECUTABLE;
+    this.exhortSyftPath = rhdaConfig.syft.executable.path || this.DEFAULT_SYFT_EXECUTABLE;
+    this.exhortSyftConfigPath = rhdaConfig.syft.config.path;
+    this.exhortSkopeoPath = rhdaConfig.skopeo.executable.path || this.DEFAULT_SKOPEO_EXECUTABLE;
+    this.exhortSkopeoConfigPath = rhdaConfig.skopeo.config.path;
+    this.exhortDockerPath = rhdaConfig.docker.executable.path || this.DEFAULT_DOCKER_EXECUTABLE;
+    this.exhortPodmanPath = rhdaConfig.podman.executable.path || this.DEFAULT_PODMAN_EXECUTABLE;
+    this.exhortImagePlatform = rhdaConfig.imagePlatform;
   }
 
   /**
@@ -80,11 +113,16 @@ class Config {
    */
   private async setProcessEnv(): Promise<void> {
     process.env['VSCEXT_STACK_ANALYSIS_COMMAND'] = this.stackAnalysisCommand;
-    process.env['VSCEXT_REDHAT_REPOSITORY_RECOMMENDATION_NOTIFICATION_COMMAND'] = this.rhRepositoryRecommendationNotificationCommand;
+    process.env['VSCEXT_TRACK_RECOMMENDATION_ACCEPTANCE_COMMAND'] = this.trackRecommendationAcceptanceCommand;
     process.env['VSCEXT_UTM_SOURCE'] = this.utmSource;
     process.env['VSCEXT_MATCH_MANIFEST_VERSIONS'] = this.matchManifestVersions;
+    process.env['VSCEXT_USE_PYTHON_VIRTUAL_ENVIRONMENT'] = this.usePythonVirtualEnvironment;
+    process.env['VSCEXT_USE_GO_MVS'] = this.useGoMVS;
+    process.env['VSCEXT_ENABLE_PYTHON_BEST_EFFORTS_INSTALLATION'] = this.enablePythonBestEffortsInstallation;
+    process.env['VSCEXT_USE_PIP_DEP_TREE'] = this.usePipDepTree;
     process.env['VSCEXT_VULNERABILITY_ALERT_SEVERITY'] = this.vulnerabilityAlertSeverity;
     process.env['VSCEXT_EXHORT_MVN_PATH'] = this.exhortMvnPath;
+    process.env['VSCEXT_EXHORT_GRADLE_PATH'] = this.exhortGradlePath;
     process.env['VSCEXT_EXHORT_NPM_PATH'] = this.exhortNpmPath;
     process.env['VSCEXT_EXHORT_GO_PATH'] = this.exhortGoPath;
     process.env['VSCEXT_EXHORT_PYTHON3_PATH'] = this.exhortPython3Path;
@@ -92,9 +130,16 @@ class Config {
     process.env['VSCEXT_EXHORT_PYTHON_PATH'] = this.exhortPythonPath;
     process.env['VSCEXT_EXHORT_PIP_PATH'] = this.exhortPipPath;
     process.env['VSCEXT_TELEMETRY_ID'] = this.telemetryId;
+    process.env['VSCEXT_EXHORT_SYFT_PATH'] = this.exhortSyftPath;
+    process.env['VSCEXT_EXHORT_SYFT_CONFIG_PATH'] = this.exhortSyftConfigPath;
+    process.env['VSCEXT_EXHORT_SKOPEO_PATH'] = this.exhortSkopeoPath;
+    process.env['VSCEXT_EXHORT_SKOPEO_CONFIG_PATH'] = this.exhortSkopeoConfigPath;
+    process.env['VSCEXT_EXHORT_DOCKER_PATH'] = this.exhortDockerPath;
+    process.env['VSCEXT_EXHORT_PODMAN_PATH'] = this.exhortPodmanPath;
+    process.env['VSCEXT_EXHORT_IMAGE_PLATFORM'] = this.exhortImagePlatform;
 
-    const token = await this.getSnykToken();
-    process.env['VSCEXT_EXHORT_SNYK_TOKEN'] = token;
+    // const token = await this.getSnykToken();
+    // process.env['VSCEXT_EXHORT_SNYK_TOKEN'] = token;
   }
 
   /**