+++ src/constants.ts
@@ -28,6 +28,10 @@ export enum Titles {
   REPORT_TITLE = `Red Hat Dependency Analytics Report`,
 }
 
+export const settingNameMappings: { [key: string]: string } = {
+  'EXHORT_PYTHON_VIRTUAL_ENV': 'Use Python Virtual Environment'
+};
+
 // Refer `name` from package.json
 export const EXTENSION_ID = 'fabric8-analytics';
 // publisher.name from package.json
@@ -44,3 +48,5 @@ export const DEFAULT_RHDA_REPORT_FILE_PATH = '/tmp/redhatDependencyAnalyticsRepo
 export const REDHAT_MAVEN_REPOSITORY = 'https://maven.repository.redhat.com/ga/';
 // Red Hat GA Repository documentation
 export const REDHAT_MAVEN_REPOSITORY_DOCUMENTATION_URL = 'https://access.redhat.com/maven-repository';
+// Red Hat certified container image catalog
+export const REDHAT_CATALOG = 'https://catalog.redhat.com/software/containers/search';