+++ test/extension.test.ts
@@ -3,7 +3,7 @@ import * as vscode from 'vscode';
 
 import * as Commands from '../src/commands';
 
-suite('Fabric8 Analytics Extension', () => {
+suite('Extension module', () => {
   test('Extension should be present', () => {
     assert.ok(vscode.extensions.getExtension('redhat.fabric8-analytics'));
   });
@@ -19,6 +19,7 @@ suite('Fabric8 Analytics Extension', () => {
     const FABRIC8_COMMANDS: string[] = [
       Commands.STACK_ANALYSIS_COMMAND,
       Commands.STACK_LOGS_COMMAND,
+      Commands.TRACK_RECOMMENDATION_ACCEPTANCE_COMMAND,
       Commands.STACK_ANALYSIS_FROM_EDITOR_COMMAND,
       Commands.STACK_ANALYSIS_FROM_EXPLORER_COMMAND,
       Commands.STACK_ANALYSIS_FROM_PIE_BTN_COMMAND,